Port Travis-CI to GitHub Actions
Travis-CI no longer has a free tier (only a free trial). That was a
major reason we used Travis-CI, so that external contributors would be
able to run the CI on their forks. The other reason was that Travis-CI
was safe to run even with untrusted code.

GitHub Actions is now reasonably safe to run untrusted code (since the
addition of the permission limiting of the github token) and it has a
free tier for external contributors. Instead of migrating to
Travis-CI.com from Travis-CI.org, let's migrate to GitHub Actions.